requirements.txt
absl-py==2.3.1
contourpy==1.3.3
cv-bridge==1.13.0.post0
cycler==0.12.1
filelock==3.19.1
fonttools==4.59.2
fsspec==2025.9.0
grpcio==1.74.0
h5py==3.14.0
hdf5plugin==5.1.0
Jinja2==3.1.6
kiwisolver==1.4.9
Markdown==3.8.2
MarkupSafe==3.0.2
matplotlib==3.10.6
mpmath==1.3.0
networkx==3.5
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
opencv-python==4.12.0.88
packaging==25.0
pandas==2.3.2
pillow==11.3.0
protobuf==6.32.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
scipy==1.16.1
setuptools==78.1.1
six==1.17.0
sympy==1.14.0
tdqm==0.0.1
tensorboard==2.20.0
tensorboard-data-server==0.7.2
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.1
triton==3.4.0
typing_extensions==4.15.0
tzdata==2025.2
Werkzeug==3.1.3
wheel==0.45.1
